fix(backend): sales_quotations 实库为 INT8 → 模型统一回 i64 并消除 sales_crm×m0044 双建表漂移
- 终检判 98663d5a 的 sales_quotation 子改动内部矛盾(部分列改 i32 留 id/sales_user_id i64)会致报价单读接口运行期 500
- 据 lib.rs:27-34 迁移序定真值: sales_crm 第3位先建 sales_quotations、execute_unprepared 不经 fix_fk_types → 实库全 BIGINT/INT8;
  m0044 第4位 CREATE IF NOT EXISTS 因表已存在= no-op
- 回退 customer_id/approved_by→i64(全表 i64 自洽): model/quotation create·update·response DTO、quotation_ops crud attach_names 取键、
  approval/convert service、customer_merge handler 同步;仅回退本表,不动同提交其它已通过表
- 根因消除: m0044 三处 sales_quotation* DDL 改成与 sales_crm 一致(BIGSERIAL/BIGINT)、注明权威在 sales_crm,新库/已建库都收敛唯一真值
- cargo fmt --check 0;残留:跨宽度 FK(BIGINT→SERIAL customers.id)PG 接受(同 custom_orders 既有模式),极低风险,如 CI migration 报 FK 再按处方 ALTER

+++ backend/src/handlers/customer_merge_handler.rs
@@ -72,9 +72,9 @@ pub async fn merge_customers(
 
     // 3. 转移销售报价
     sales_quotation::Entity::update_many()
-        .filter(sales_quotation::Column::CustomerId.eq(req.source_customer_id))
+        .filter(sales_quotation::Column::CustomerId.eq(req.source_customer_id as i64))
         .set(sales_quotation::ActiveModel {
-            customer_id: Set(req.target_customer_id),
+            customer_id: Set(req.target_customer_id as i64),
             ..Default::default()
         })
         .exec(&txn)
